Demonstrate support for exclusions in BOM import
This change demonstrates how support for <exclusions> defined in
an imported BOM could be used to allow a BOM to:
1. Take opinions from upstream BOM (<scope>import</scope>)
2. Override opinions in upstream BOM (<exclusions/> + replacement <dependency/>)
3. Exclude opinions in upstream BOM (<exclusions/>

Note that the Gradle version referenced by the wrapper was published
from the `dd/dm/bom-excludes` branch. These changes have not yet
been merged into `master`. There are known issues with this codebase
and it should be used for demonstration purposes only.